+++ app/Http/Controllers/PostController.php
@@ -15,11 +15,11 @@ class PostController extends Controller
         ]); 
     }
 
-    public function show($slug)
+    public function show(Post $post)
     {
         return view('post', [
             "title" => "Single Post",
-            "post" => Post::find($slug)
+            "post" => $post
         ]);
     }
 }
