+++ resources/views/post.blade.php
@@ -1,11 +1,10 @@
 @extends('layouts.main')
 
 @section('container')
-<article>
-    <h2>{{ $post["title"] }}</h2>
-    <h5>{{ $post["author"] }}</h5>
-    <p>{{ $post["body"] }}</p>
-</article>
+    <h1 class="mb-5">{{ $post->title }}</h2>
 
-<a href="/posts">Back to post</a>
+    {!! $post->body !!}
+
+    <a href="/posts">Back to post</a>
+
 @endsection
