+++ resources/views/posts.blade.php
@@ -1,14 +1,14 @@
 @extends('layouts.main')
 
 @section('container')
+    <h1 class="mb-5">Halaman Blog Posts</h1>
 
     @foreach ($posts as $post)
-        <article class="mb-5">
+        <article class="mb-4">
             <h2>
-                <a href="/posts/{{ $post["slug"] }}">{{ $post["title"] }}</a>
+                <a href="/posts/{{ $post->slug }}">{{ $post->title }}</a>
             </h2>
-            <h5>By: {{ $post["author"] }}</h5>
-            <p>{{ $post["body"] }}</p>
+            <p>{{ $post->excerpt }}</p>
         </article>
     @endforeach
